+++ splash/qtrender_lua.py
@@ -115,13 +115,22 @@ def command(table_argument=False, sets_callback=False,
                 emits_lua_objects(meth)
             ),
             error_as_flag,
-            result_as_flag
+            result_as_flag,
         )
         meth._is_command = True
         meth._sets_callback = sets_callback
         return meth
 
     return decorator
+
+
+def _command_result_to_pyresult(res):
+    """ Convert result of a method decorated with @command to PyResult.
+    Use this method to call methods decorated with @command inside other
+    methods decorated with @command.
+    """
+    op, cmd = res
+    return PyResult(cmd, _operation=to_unicode(op))
 
 
 def lua_property(name):
@@ -233,13 +242,13 @@ def add_flag(tuple, flag):
     return new_tuple
 
 
-def exceptions_as_return_values(meth, error_as_flag=False, result_as_flag=False):
+def exceptions_as_return_values(meth, error_as_flag=False,
+                                result_as_flag=False):
     """Decorator for allowing Python exceptions to be caught from Lua.
 
     TODO: this decorator is the last one on the way from Python to Lua and thus
     is responsible for converting non-PyResult values to PyResult.  This is
     suboptimal and should be fixed.
-
     """
     @functools.wraps(meth)
     def exceptions_as_return_values_wrapper(self, *args, **kwargs):
@@ -257,7 +266,6 @@ def exceptions_as_return_values(meth, error_as_flag=False, result_as_flag=False)
                 res = (b'return', False, repr(e).encode('utf-8'))
             else:
                 res = (b'raise', repr(e).encode('utf-8'))
-
         return res
 
     return exceptions_as_return_values_wrapper
@@ -942,6 +950,29 @@ class Splash(BaseExposedObject):
     def send_text(self, text):
         self.tab.send_text(text)
 
+    @lua_property('scroll_position')
+    @command()
+    def get_scroll_position(self):
+        return self.tab.get_scroll_position()
+
+    @get_scroll_position.lua_setter
+    @command()
+    def scroll_to(self, x=None, y=None):
+        if x is None or y is None:
+            pos = self.tab.get_scroll_position()
+            x = pos['x'] if x is None else x
+            y = pos['y'] if y is None else y
+
+        for value, name in [(x, "x"), (y, "y")]:
+            if not isinstance(value, (int, float)):
+                raise ScriptError({
+                    "argument": name,
+                    "message": "scroll {} coordinate must be "
+                               "a number, got {}".format(name, repr(value))
+                })
+
+        self.tab.set_scroll_position(x, y)
+
     @command()
     def set_content(self, data, mime_type=None, baseurl=None):
         if isinstance(data, str):
@@ -1315,6 +1346,7 @@ class Splash(BaseExposedObject):
                 sandboxed=False,
                 strict=self.strict_lua_runner,
             )
+            self._objects_to_clear.add(runner)
             coro = self.lua.create_coroutine(callback)
             runner.start(coro, coro_args, return_result, return_error)
             return runner
@@ -1333,6 +1365,9 @@ class Splash(BaseExposedObject):
             lambda o: isinstance(o, HTMLElement),
             convert=_expose,
         )
+
+    def wait_tick(self, time=0.0):
+        return _command_result_to_pyresult(self.wait(time))
 
 
 class _ExposedTimer(BaseExposedObject):
@@ -1467,9 +1502,12 @@ class _ExposedElement(BaseExposedObject):
         'requestFullscreen',
         'requestPointerLock',
         'scrollIntoView',
+        'scrollIntoViewIfNeeded',
         'setAttribute',
         'setAttributeNS',
-        'setPointerCapture'
+        'setPointerCapture',
+        'scrollIntoView',
+        'scrollIntoViewIfNeeded',
     ]
     NODE_METHODS = [
         'appendChild',
@@ -1734,6 +1772,7 @@ class _ExposedElement(BaseExposedObject):
             })
 
         self.element.mouse_click(x, y)
+        return self.splash.wait_tick()
 
     @command(error_as_flag=True)
     def mouse_hover(self, x=None, y=None):
@@ -1752,6 +1791,7 @@ class _ExposedElement(BaseExposedObject):
             })
 
         self.element.mouse_hover(x, y)
+        return self.splash.wait_tick()
 
     @command()
     def styles(self):
@@ -2165,6 +2205,7 @@ class SplashCoroutineRunner(BaseScriptRunner):
     """
     def __init__(self, lua, splash, log, sandboxed, strict):
         self.splash = splash
+        self._exited = False
         super(SplashCoroutineRunner, self).__init__(
             lua=lua,
             log=log,
@@ -2172,21 +2213,29 @@ class SplashCoroutineRunner(BaseScriptRunner):
             strict=strict,
         )
 
-    def start(self, coro_func, coro_args=None, return_result=None, return_error=None):
+    def start(self, coro_func, coro_args=None, return_result=None,
+              return_error=None):
         do_nothing = lambda *args, **kwargs: None
         self.return_result = return_result or do_nothing
         self.return_error = return_error or do_nothing
         super(SplashCoroutineRunner, self).start(coro_func, coro_args)
 
     def on_result(self, result):
+        if self._exited:
+            return
         self.return_result(result)
 
     def on_async_command(self, cmd):
+        if self._exited:
+            return
         self.splash.run_async_command(cmd)
 
     @stop_on_error
     def dispatch(self, cmd_id, *args):
         super(SplashCoroutineRunner, self).dispatch(cmd_id, *args)
+
+    def clear(self):
+        self._exited = True
 
 
 class MainCoroutineRunner(SplashCoroutineRunner):
@@ -2219,6 +2268,8 @@ class MainCoroutineRunner(SplashCoroutineRunner):
         # would be serialized like that anyway.
         #
         # FIXME: maybe request writer must be fixed?
+        if self._exited:
+            return
         if isinstance(result, tuple):
             result = list(result)
         self.return_result((
@@ -2229,6 +2280,8 @@ class MainCoroutineRunner(SplashCoroutineRunner):
         ))
 
     def on_lua_error(self, lua_exception):
+        if self._exited:
+            return
         py_exception = self.exceptions.get_last()
 
         if not py_exception: